Open multiple images, clean dict on search
When opening multiple images from File -> Open, it will place them in a
collection to be exclusively viewed. When searching for images that do
not have a generated icon, notify the user we're preparing the preview.
Before generating any icons, we need to make sure that the dictionary is
clean of any non-existant files.