Zoom at center of image, add drag-and-drop
Zooming in/out will set the scrollbar position to the center of the
image. Make certain the tag text-field refreshes so it will accurately
display our tag changes after we've made them. Add drag-and-drop feature
such that it appends to the current collection. Make sure that appending
files to a collection does not result in duplicate paths (weird things
happen if this occurs).