Cache embedded XMP, more work on meta-editor
When loading the image, cache the XMP metadata so we're not slicing up
the bytes quote so often. Also, allow the meta-editor to modify attrs
inside rdf:Description. Still thinking about whether or not I should
remove xmp:Identifier, since it seems to be a little more tricky than
the others.